+++ tests/System/CheckDirectDependencyUseCommandTest.php
@@ -45,6 +45,7 @@ class CheckDirectDependencyUseCommandTest extends SystemTestCase
             ],
             'Matomo\Network' => [
                 'TrackingSpamPrevention/AllowListIpRange.php',
+                'TrackingSpamPrevention/BlockListIpRange.php',
                 'TrackingSpamPrevention/BlockedIpRanges.php',
                 'TrackingSpamPrevention/TrackingSpamPrevention.php',
             ],